refactor(condo): DOMA-7397 Add statuses file
