+++ apps/condo/domains/news/constants/newsItemSharingStatuses.js
@@ -1,0 +1,18 @@
+const STATUSES = {
+    SCHEDULED: 'scheduled',
+    PROCESSING: 'processing',
+    PUBLISHED: 'published',
+    ERROR: 'error',
+}
+
+const ALLOWED_TRANSITIONS = {
+    [STATUSES.SCHEDULED]: [STATUSES.PROCESSING, STATUSES.ERROR],
+    [STATUSES.PROCESSING]: [STATUSES.PUBLISHED, STATUSES.ERROR],
+    [STATUSES.PUBLISHED]: [STATUSES.ERROR],
+    [STATUSES.ERROR]: [],
+}
+
+module.exports = {
+    STATUSES,
+    ALLOWED_TRANSITIONS,
+}
